@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.0-144.0.dev <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -39,6 +39,9 @@ dependencies:
   equatable: ^2.0.3
   bloc: ^8.0.3
   flutter_bloc: ^8.0.1
+  fl_chart: ^0.51.0
+  intl: ^0.17.0
+  timeago: ^3.2.2
 
 dev_dependencies:
   flutter_test:
