@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
 
-import '../api_serivce.dart';
-import '../model/model.dart';
+import '../api_service.dart';
+import '../model/_model.dart';
 
 part 'coin_provider.dart';
