@@ -2,17 +2,18 @@ import 'package:device_preview/device_preview.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'data/provider/provider.dart';
-import 'screen/tab_navigation/coin/coin_bloc.dart';
-import 'screen/tab_navigation/tab_navigation_screen.dart';
+import 'module/coin_detail/bloc/coin_detail_bloc.dart';
+import 'module/home/coin/coin_bloc.dart';
+import 'module/home/explore/explore_bloc.dart';
+import 'module/home/home_screen.dart';
 import 'theme/app_theme.dart';
 
 void main() {
   runApp(MultiBlocProvider(
     providers: [
-      BlocProvider<CoinBloc>(
-        create: (context) => CoinBloc(coinProvider: CoinProvider()),
-      )
+      BlocProvider<CoinBloc>(create: (_) => CoinBloc()),
+      BlocProvider(create: (_) => CoinDetailBloc()),
+      BlocProvider(create: (_) => ExploreBloc()),
     ],
     child: DevicePreview(
       enabled: false,
@@ -24,7 +25,7 @@ void main() {
         title: 'Flutter Demo',
         theme: AppTheme.light,
         darkTheme: AppTheme.dark,
-        home: const TabNavigationScreen(),
+        home: const HomeScreen(),
       ),
     ),
   ));
