@@ -1,8 +1,4 @@
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/cupertino.dart';
-import 'package:flutter_svg/svg.dart';
-
-import '../theme/app_color.dart';
+part of '_widget.dart';
 
 class WidgetImage extends StatelessWidget {
   final String url;
@@ -10,6 +6,7 @@ class WidgetImage extends StatelessWidget {
   final double? width;
   final bool isSvg;
   final BoxFit fit;
+  final double border;
 
   const WidgetImage({
     Key? key,
@@ -18,13 +15,16 @@ class WidgetImage extends StatelessWidget {
     this.width,
     this.isSvg = false,
     this.fit = BoxFit.cover,
+    this.border = 0,
   }) : super(key: key);
 
   Widget buildPlaceHolder(BuildContext context) {
     return Container(
       alignment: Alignment.center,
       decoration: BoxDecoration(
-          color: AppColors.tertiary, borderRadius: BorderRadius.circular(10)),
+        color: AppColors.tertiary,
+        borderRadius: BorderRadius.circular(10),
+      ),
       child: const CupertinoActivityIndicator(
         color: AppColors.primary,
         radius: 4,
@@ -37,11 +37,15 @@ class WidgetImage extends StatelessWidget {
     if (isSvg) {
       return SvgPicture.network(url, placeholderBuilder: buildPlaceHolder);
     }
-    return CachedNetworkImage(
-      imageUrl: url,
-      height: height,
-      width: width,
-      placeholder: (context, s) => buildPlaceHolder(context),
+    return ClipRRect(
+      borderRadius: BorderRadius.circular(border),
+      child: CachedNetworkImage(
+        imageUrl: url,
+        height: height,
+        width: width,
+        fit: fit,
+        placeholder: (context, s) => buildPlaceHolder(context),
+      ),
     );
   }
 }
